user-story-tests/23-pass-through-from-design-system.html

<!DOCTYPE html>
<html lang="en">

<head>
    <meta charset="UTF-8">
    <meta name="viewport" content="width=device-width, initial-scale=1.0">
    <title>Pass through from design system</title>
    <script src='./resources/testharness.js'></script>
    <script src='./resources/testharnessreport.js'></script>
</head>

<body>

    <style>
        button {
            border: thick solid black;
        }

        @layer design-system {
            button {
                border: thick dashed red;
            }
        }
    </style>
    <button>Button outside a shadow tree (thick solid black)</button>
    <button-group>
        <template shadowrootmode="open">
            <style>
                @layer inherit.design-system.as.design-system, design-system;
            </style>
            <button>
                Button inside a shadow tree (thick dashed red) (from inherit.design-system.as.design-system)
            </button>
        </template>
    </button-group>

    <script type="module">
        import { adoptStylesFromLayerStatementRule } from "../AdoptStyles.js";

        window.onload = function () {

            const elName = "button-group"

            test(function () {

                const inheritResult = adoptStylesFromLayerStatementRule(document.querySelector(elName))
                assert_true(inheritResult, "inherit found, so returns true");
            }, "Pass through from design system");

        }
    </script>

    <div id="log"></div>

</body>

</html>
